Update project_restaurant_page.md
I had a very hard time figuring out the .gitignore file. For whatever reason, even after following multiple online sources, I was only able to make it work when I did it globally for all future repos. Maybe the lack of information is intentional so that one has to figure this out themselves, in which case this request can be ignored. I definitely respect that philosophy, but if not a resource or two on setting up the .gitignore file may be helpful.